Good progress writing unit tests.  Done most files in TDL but still need to do TdlRule and RuleRunner. Still need to do RuleWriter. Made a lot of breaking changes.  The code under test definitely does't compiler any more.

+++ tests/mocks/time_test/mktime.c
@@ -35,7 +35,7 @@
 
 #include <time_test.h>
 
-extern long     __utc_offset;
+extern int32_t     __utc_offset;
 
 extern int      (*__dst_ptr) (const time_t_test *, int32_t *);
